qase.io integration moved all helpers/util to main/java
