+++ kafka-ui-e2e-checks/pom.xml
@@ -14,25 +14,26 @@
         <kafka-ui-contract>${project.version}</kafka-ui-contract>
         <junit.version>5.8.2</junit.version>
         <aspectj.version>1.9.8</aspectj.version>
-        <allure.version>2.17.2</allure.version>
-        <json-smart.version>1.3.3</json-smart.version>
-        <testcontainers.version>1.16.3</testcontainers.version>
-        <selenide.version>6.3.0</selenide.version>
-        <assertj.version>3.17.1</assertj.version>
-        <google.auto-service.version>1.0-rc7</google.auto-service.version>
+        <allure.version>2.18.1</allure.version>
+        <json-smart.version>2.4.8</json-smart.version>
+        <testcontainers.version>1.17.2</testcontainers.version>
+        <selenide.version>6.6.3</selenide.version>
+        <assertj.version>3.23.1</assertj.version>
+        <google.auto-service.version>1.0.1</google.auto-service.version>
         <hamcrest.version>2.2</hamcrest.version>
-        <slf4j.version>1.7.32</slf4j.version>
-        <testcontainers.junit-jupiter.version>1.16.3</testcontainers.junit-jupiter.version>
-        <allure.java-commons.version>2.17.2</allure.java-commons.version>
-        <dotenv.version>2.2.0</dotenv.version>
+        <slf4j.version>1.7.36</slf4j.version>
+        <testcontainers.junit-jupiter.version>1.17.2</testcontainers.junit-jupiter.version>
+        <allure.java-commons.version>2.18.1</allure.java-commons.version>
+        <dotenv.version>2.2.4</dotenv.version>
         <junit.platform-launcher.version>1.8.2</junit.platform-launcher.version>
         <allure.maven-plugin.version>2.6</allure.maven-plugin.version>
         <ashot.version>1.5.4</ashot.version>
-        <allure.screendiff-plugin.version>2.17.2</allure.screendiff-plugin.version>
+        <allure.screendiff-plugin.version>2.18.1</allure.screendiff-plugin.version>
         <maven.surefire-plugin.version>2.22.2</maven.surefire-plugin.version>
         <allure-maven.version>2.10.0</allure-maven.version>
         <kafka.version>3.0.0</kafka.version>
         <netty.version>4.1.77.Final</netty.version>
+        <qase.io.version>2.1.3</qase.io.version>
     </properties>
 
     <dependencies>
@@ -207,7 +208,7 @@
         <dependency>
             <groupId>org.junit.platform</groupId>
             <artifactId>junit-platform-launcher</artifactId>
-            <version>${junit.platform-launcher.version}</version>
+           <version>${junit.platform-launcher.version}</version>
         </dependency>
         <dependency>
             <groupId>ru.yandex.qatools.allure</groupId>
@@ -234,13 +235,22 @@
             <groupId>com.provectus</groupId>
             <artifactId>kafka-ui-contract</artifactId>
             <version>${kafka-ui-contract}</version>
-            <scope>test</scope>
         </dependency>
         <dependency>
             <groupId>org.testcontainers</groupId>
             <artifactId>selenium</artifactId>
             <version>${testcontainers.version}</version>
             <scope>test</scope>
+        </dependency>
+            <dependency>
+                <groupId>io.qase</groupId>
+                <artifactId>qase-junit5</artifactId>
+                <version>${qase.io.version}</version>
+            </dependency>
+        <dependency>
+            <groupId>io.qase</groupId>
+            <artifactId>qase-api</artifactId>
+            <version>${qase.io.version}</version>
         </dependency>
     </dependencies>
 
@@ -258,7 +268,7 @@
                         <artifactId>maven-surefire-plugin</artifactId>
                         <version>${maven.surefire-plugin.version}</version>
                         <configuration>
-                            <skipTests>true</skipTests>
+                            <skipTests>false</skipTests>
                         </configuration>
                     </plugin>
                     <plugin>
