+++ kafka-ui-e2e-checks/src/main/java/helpers/Helpers.java
@@ -1,0 +1,11 @@
+package helpers;
+
+
+
+public class Helpers {
+    public static final Helpers INSTANCE = new Helpers();
+
+    private Helpers(){}
+
+    public ApiHelper apiHelper = new ApiHelper();
+}
