--- kafka-ui-e2e-checks/src/test/java/com/provectus/kafka/ui/helpers/Helpers.java
@@ -1,11 +1,0 @@
-package com.provectus.kafka.ui.helpers;
-
-
-
-public class Helpers {
-    public static final Helpers INSTANCE = new Helpers();
-
-    private Helpers(){}
-
-    public ApiHelper apiHelper = new ApiHelper();
-}
